+++ mobile/kmp/sample/src/commonMain/kotlin/com/atls/hyperion/sample/App.kt
@@ -25,6 +25,8 @@ import com.atls.hyperion.ui.fragments.datepicker.stories.DateRangePickerStory
 import com.atls.hyperion.ui.primitives.stories.LinkStory
 import com.atls.hyperion.ui.primitives.stories.TextStory
 import com.atls.hyperion.ui.theme.Theme
+import com.atls.hyperion.ui.theme.tokens.effects.stories.opacity.OpacityStory
+import com.atls.hyperion.ui.theme.tokens.elevation.stories.ElevationStory
 
 @Composable
 fun App() {
@@ -41,7 +43,9 @@ fun App() {
                 DatePickerStory(),
                 DateRangePickerStory(),
                 DividerStory(),
+                ElevationStory(),
                 InputStory(),
+                OpacityStory(),
                 PaginationStory(),
                 PlaceholderStory(),
                 PopupStory(),