+++ mobile/kmp/ui/src/commonMain/kotlin/com/atls/hyperion/ui/components/button/styles/shape/Constants.kt
@@ -1,0 +1,10 @@
+package com.atls.hyperion.ui.components.button.styles.shape
+
+import androidx.compose.ui.unit.dp
+
+internal object MinHeight {
+    val xs = 24.dp
+    val sm = 28.dp
+    val md = 32.dp
+    val lg = 32.dp
+}