+++ ui-parts/design/src/main.ts
@@ -16,8 +16,8 @@ const config: StorybookConfig = {
   stories: [
     '../../**/src/*.stories.@(js|jsx|mjs|ts|tsx|mdx)',
     '../../**/src/*.mdx',
-    '../../**/stories/*.stories.@(js|jsx|mjs|ts|tsx|mdx)',
-    '../../**/stories/*.mdx',
+    '../../**/stories/**/*.stories.@(js|jsx|mjs|ts|tsx|mdx)',
+    '../../**/stories/**/*.mdx',
   ],
   addons: [
     getAbsolutePath('@storybook/addon-webpack5-compiler-swc'),