+++ ui-admin/upload/CHANGELOG.md
@@ -1,5 +1,19 @@
 
 
+# [1.2.0](https://github.com/atls/hyperion/compare/@atls-ui-admin/upload@1.2.0...@atls-ui-admin/upload@1.2.0) (2026-08-26)
+
+
+
+
+
+
+# [1.2.0](https://github.com/atls/hyperion/compare/@atls-ui-admin/upload@1.2.0...@atls-ui-admin/upload@1.2.0) (2026-08-25)
+
+
+
+
+
+
 # [1.2.0](https://github.com/atls/hyperion/compare/@atls-ui-admin/upload@1.2.0...@atls-ui-admin/upload@1.2.0) (2026-08-18)