+++ ui-parts/carousel/CHANGELOG.md
@@ -1,5 +1,19 @@
 
 
+# [1.1.0](https://github.com/atls/hyperion/compare/@atls-ui-parts/carousel@1.1.0...@atls-ui-parts/carousel@1.1.0) (2026-08-26)
+
+
+
+
+
+
+# [1.1.0](https://github.com/atls/hyperion/compare/@atls-ui-parts/carousel@1.1.0...@atls-ui-parts/carousel@1.1.0) (2026-08-25)
+
+
+
+
+
+
 # [1.1.0](https://github.com/atls/hyperion/compare/@atls-ui-parts/carousel@1.1.0...@atls-ui-parts/carousel@1.1.0) (2026-08-18)